Former-commit-id: e2de5d8d0e95b8f6b8c035a67da5c761ded11d29

+++ moduleapi/src/main/java/com/simprints/moduleapi/face/responses/IFaceMatchResult.kt
@@ -4,5 +4,5 @@ import android.os.Parcelable
 
 interface IFaceMatchResult: Parcelable {
     val guid: String
-    val confidence: Int
+    val confidence: Float
 }
